fix "[Object] object" that shows up on ortholog page changes.
the problem is the intermittent updates for the watch query are sometimes improperly formatted. I think it must be a problem in the graphql library though.

Probably a better fix for us would be to fetch only what's needed for this query. It also takes too long to switch pages on orthologs. Or maybe we can fetch all the orthologs and page them by slicing it up, like we do for expression data.